Add two entrance points for heroku
