+++ startbot.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 chmod +x engines/stockfish_14.1_linux_x64_popcnt
 chmod a+x engines/fsf
-python user_interface.py -u
+python heroku_challenges.py
